+++ commands/spigot-platform/src/main/java/net/apartium/cocoabeans/commands/spigot/requirements/factory/WhitelistRequirementFactory.java
@@ -11,6 +11,7 @@
 package net.apartium.cocoabeans.commands.spigot.requirements.factory;
 
 import net.apartium.cocoabeans.commands.CommandNode;
+import net.apartium.cocoabeans.commands.GenericNode;
 import net.apartium.cocoabeans.commands.requirements.*;
 import net.apartium.cocoabeans.commands.spigot.exception.WhitelistException;
 import net.apartium.cocoabeans.commands.spigot.requirements.Whitelist;
@@ -30,7 +31,7 @@ public class WhitelistRequirementFactory implements RequirementFactory {
 
     @Nullable
     @Override
-    public Requirement getRequirement(CommandNode node, Object obj) {
+    public Requirement getRequirement(GenericNode node, Object obj) {
         if (!(obj instanceof Whitelist whitelist))
             return null;