+++ commands/spigot-platform/src/test/java/net/apartium/cocoabeans/commands/spigot/parsers/Meow.java
@@ -1,4 +1,4 @@
-package net.apartium.cocoabeans.commands.parser;
+package net.apartium.cocoabeans.commands.spigot.parsers;
 
 public record Meow(String cat, int age, Gender gender) {
 
@@ -6,4 +6,8 @@ public record Meow(String cat, int age, Gender gender) {
         MALE, FEMALE, OTHER
     }
 
+    @Override
+    public String toString() {
+        return "cat: " + cat + ", age: " + age + ", gender: " + gender;
+    }
 }