+++ commands/src/main/java/net/apartium/cocoabeans/commands/RegisteredVariant.java
@@ -11,27 +11,86 @@
 package net.apartium.cocoabeans.commands;
 
 import net.apartium.cocoabeans.commands.requirements.ArgumentRequirement;
+import net.apartium.cocoabeans.commands.requirements.ArgumentRequirementFactory;
+import org.jetbrains.annotations.ApiStatus;
 
 import java.lang.invoke.MethodHandle;
+import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
-public record RegisteredCommandVariant(
+/**
+ * Registered variant for a command
+ * @param method target method to run
+ * @param parameters parameters
+ * @param node node instance
+ * @param argumentIndexList list of argument indexes that will be passed to the method
+ * @param priority the priority of the variant
+ */
+public record RegisteredVariant(
         MethodHandle method,
         Parameter[] parameters,
-        CommandNode commandNode,
+        GenericNode node,
         List<ArgumentIndex<?>> argumentIndexList,
         int priority
 ) {
 
+    /**
+     * Comparator for registered variants
+     */
+    public static final Comparator<RegisteredVariant> REGISTERED_VARIANT_COMPARATOR = (a, b) -> Integer.compare(b.priority(), a.priority());
+
+    /**
+     * Parameter represent a parameter of the command after has been parsed
+     * @param type The type of parameter
+     * @param parameterizedType parameterized type
+     * @param argumentRequirements argument requirements
+     * @param parameterName parameter name
+     */
     public record Parameter(
             Class<?> type,
             Type parameterizedType,
             ArgumentRequirement[] argumentRequirements,
             String parameterName
     ) {
+
+        /**
+         * create parameters from the given parameters
+         * @param node instance of the node
+         * @param parameters parameters
+         * @param argumentRequirementFactories argument requirement factories for caching
+         * @return parameters parsed
+         */
+        @ApiStatus.AvailableSince("0.0.37")
+        public static Parameter[] of(GenericNode node, java.lang.reflect.Parameter[] parameters, Map<Class<? extends ArgumentRequirementFactory>, ArgumentRequirementFactory> argumentRequirementFactories) {
+            Parameter[] result = new Parameter[parameters.length];
+
+            for (int i = 0; i < result.length; i++) {
+                result[i] = new Parameter(
+                        parameters[i].getType(),
+                        parameters[i].getParameterizedType(),
+                        ArgumentRequirementFactory.createArgumentRequirements(node, parameters[i].getAnnotations(), argumentRequirementFactories),
+                        getParamName(parameters[i])
+                );
+            }
+
+            return result;
+        }
+
+        @ApiStatus.Internal
+        private static String getParamName(java.lang.reflect.Parameter parameter) {
+            String name = Optional.ofNullable(parameter.getAnnotation(Param.class))
+                    .map(Param::value)
+                    .orElse(null);
+
+            if (name == null)
+                return null;
+
+            if (name.isEmpty())
+                throw new IllegalArgumentException("Parameter name cannot be empty");
+
+            return name;
+        }
 
         @Override
         public boolean equals(Object o) {