+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/CompoundParserOption.java
@@ -2,17 +2,28 @@ package net.apartium.cocoabeans.commands.parsers;
 
 import net.apartium.cocoabeans.commands.AbstractCommandProcessingContext;
 import net.apartium.cocoabeans.commands.CommandProcessingContext;
+import net.apartium.cocoabeans.commands.RegisterArgumentParser;
+import net.apartium.cocoabeans.commands.RegisteredVariant;
 import net.apartium.cocoabeans.structs.Entry;
 
 import java.util.*;
 
 /* package-private */ class CompoundParserOption<T> {
 
-    final List<Entry<ArgumentParser<?>, CompoundParserBranchProcessor<T>>> argumentTypeHandlerMap = new ArrayList<>();
+    private final List<RegisteredVariant> registeredVariants = new ArrayList<>();
+
+    final List<Entry<RegisterArgumentParser<?>, CompoundParserBranchProcessor<T>>> argumentTypeHandlerMap = new ArrayList<>();
 
 
     public Optional<CompoundParser.ParserResult> parse(CommandProcessingContext processingContext) {
-        for (Entry<ArgumentParser<?>, CompoundParserBranchProcessor<T>> entry : argumentTypeHandlerMap) {
+        if (processingContext.args().size() <= processingContext.index() || argumentTypeHandlerMap.isEmpty()) {
+            if (!registeredVariants.isEmpty())
+                return Optional.of(new CompoundParser.ParserResult(Collections.unmodifiableList(registeredVariants), processingContext.index(), new HashMap<>()));
+
+            return Optional.empty();
+        }
+
+        for (Entry<RegisterArgumentParser<?>, CompoundParserBranchProcessor<T>> entry : argumentTypeHandlerMap) {
             Optional<? extends ArgumentParser.ParseResult<?>> parse = entry.key().parse(processingContext);
 
             if (parse.isEmpty())
@@ -35,38 +46,19 @@ import java.util.*;
             return result;
         }
 
+        if (!registeredVariants.isEmpty())
+            return Optional.of(new CompoundParser.ParserResult(Collections.unmodifiableList(registeredVariants), processingContext.index(), new HashMap<>()));
+
         return Optional.empty();
     }
 
-    public OptionalInt tryParse(CommandProcessingContext processingContext) {
-        for (Entry<ArgumentParser<?>, CompoundParserBranchProcessor<T>> entry : argumentTypeHandlerMap) {
-            OptionalInt parse = entry.key().tryParse(processingContext);
-
-            if (parse.isEmpty())
-                continue;
-
-            OptionalInt result = entry.value().tryParse(new AbstractCommandProcessingContext(
-                    processingContext.sender(),
-                    processingContext.label(),
-                    processingContext.args().toArray(new String[0]),
-                    parse.getAsInt()
-            ));
-
-            if (result.isEmpty())
-                continue;
-
-            return result;
-        }
-
-        return OptionalInt.empty();
-    }
 
     public Optional<ArgumentParser.TabCompletionResult> tabCompletion(CommandProcessingContext processingContext) {
         Set<String> result = new HashSet<>();
 
         int highestIndex = -1;
 
-        for (Entry<ArgumentParser<?>, CompoundParserBranchProcessor<T>> entry : argumentTypeHandlerMap) {
+        for (Entry<RegisterArgumentParser<?>, CompoundParserBranchProcessor<T>> entry : argumentTypeHandlerMap) {
             Optional<ArgumentParser.TabCompletionResult> parse = entry.key().tabCompletion(processingContext);
 
             if (parse.isEmpty())
@@ -81,6 +73,8 @@ import java.util.*;
 
         return Optional.of(new ArgumentParser.TabCompletionResult(result, highestIndex));
     }
+
+    public List<RegisteredVariant> getRegisteredVariants() { return registeredVariants; }
 
 
 }