+++ commands/src/main/java/net/apartium/cocoabeans/commands/requirements/argument/RangeArgumentRequirementFactory.java
@@ -12,6 +12,7 @@ package net.apartium.cocoabeans.commands.requirements.argument;
 
 import net.apartium.cocoabeans.commands.CommandContext;
 import net.apartium.cocoabeans.commands.CommandNode;
+import net.apartium.cocoabeans.commands.GenericNode;
 import net.apartium.cocoabeans.commands.Sender;
 import net.apartium.cocoabeans.commands.requirements.ArgumentRequirement;
 import net.apartium.cocoabeans.commands.requirements.ArgumentRequirementFactory;
@@ -21,7 +22,7 @@ import org.jetbrains.annotations.Nullable;
 public class RangeArgumentRequirementFactory implements ArgumentRequirementFactory {
 
     @Override
-    public @Nullable ArgumentRequirement getArgumentRequirement(CommandNode commandNode, Object obj) {
+    public @Nullable ArgumentRequirement getArgumentRequirement(GenericNode node, Object obj) {
         if (obj == null)
             return null;